Revert r3419, r3420,r3424 as they were checked into 1.4.4 tag instead of 1.4-stable branch. The tag should remain as it was when 1.4.4 was released

+++ master/lib/Munin/Master/ProcessManager.pm
@@ -85,8 +85,11 @@ sub start_work {
 sub _prepare_unix_socket {
     my ($self) = @_;
 
-    unlink $self->{socket_file}
-        or $! ne 'No such file or directory' && croak "unlink failed: $!";
+    if ( -e $self->{socket_file} ) {
+	    unlink $self->{socket_file}
+        	or croak "unlink failed: $!";
+    }
+
     socket my $sock, PF_UNIX, SOCK_STREAM, 0
         or croak "socket failed: $!";
     bind $sock, sockaddr_un($self->{socket_file})
